moving context provider to directory layout
